share handleSurveyDataTransferRequest between inline cta and audienceactionflow (#3727)
* share handleSurveyDataTransferRequest between inline cta and audience action flow

* update tests

* remove unused code

* delete unused code

* add survey-utils

* Add copyright

* remove unused code
